requirements.txt
torchaudio
imageio[ffmpeg]